Reactivate test_smirnoff.py tests and move TestParametersList to test_parameters.py
