+++ openforcefield/tests/test_parameters.py
@@ -1,0 +1,88 @@
+#!/usr/bin/env python
+
+#=============================================================================================
+# MODULE DOCSTRING
+#=============================================================================================
+
+"""
+Test classes and function in module openforcefield.typing.engines.smirnoff.parameters.
+
+"""
+
+
+#=============================================================================================
+# GLOBAL IMPORTS
+#=============================================================================================
+
+
+from openforcefield.typing.engines.smirnoff.parameters import ParameterList, ParameterType
+
+
+#=============================================================================================
+# PARAMETER LIST
+#=============================================================================================
+
+
+class TestParameterList:
+    """Test capabilities of ParameterList for accessing and manipulating SMIRNOFF parameter definitions.
+    """
+
+    def test_create(self):
+        """Test creation of a parameter list.
+        """
+        p1 = ParameterType(smirks='[*:1]')
+        p2 = ParameterType(smirks='[#1:1]')
+        parameters = ParameterList([p1, p2])
+
+    def test_getitem(self):
+        """Test ParameterList __getitem__ overloading.
+        """
+        p1 = ParameterType(smirks='[*:1]')
+        p2 = ParameterType(smirks='[#1:1]')
+        parameters = ParameterList([p1, p2])
+        assert parameters[0] == p1
+        assert parameters[1] == p2
+        assert parameters[p1.smirks] == p1
+        assert parameters[p2.smirks] == p2
+
+    def test_setitem(self):
+        """Test ParameterList __setitem__ overloading.
+        """
+        p1 = ParameterType(smirks='[*:1]')
+        p2 = ParameterType(smirks='[#1:1]')
+        p3 = ParameterType(smirks='[#7:1]')
+        parameters = ParameterList([p1, p2, p3])
+        parameters[0].smirks = '[*X4:1]'
+        assert p1.smirks == '[*X4:1]'
+        parameters['[*X4:1]'].smirks = '[*:1]'
+        assert p1.smirks == '[*:1]'
+
+    def test_contains(self):
+        """Test ParameterList __contains__ overloading.
+        """
+        p1 = ParameterType(smirks='[*:1]')
+        p2 = ParameterType(smirks='[#1:1]')
+        p3 = ParameterType(smirks='[#7:1]')
+        parameters = ParameterList([p1, p2])
+        assert p1 in parameters
+        assert p2 in parameters
+        assert p3 not in parameters
+        assert p1.smirks in parameters
+        assert p2.smirks in parameters
+        assert p3.smirks not in parameters
+
+    def test_del(self):
+        """Test ParameterList __del__ overloading.
+        """
+        p1 = ParameterType(smirks='[*:1]')
+        p2 = ParameterType(smirks='[#1:1]')
+        p3 = ParameterType(smirks='[#7:1]')
+        parameters = ParameterList([p1, p2, p3])
+        del parameters[2]
+        assert len(parameters) == 2
+        assert p1 in parameters
+        assert p2 in parameters
+        assert p3 not in parameters
+        del parameters['[#1:1]']
+        assert p1 in parameters
+        assert p2 not in parameters
